pyproject.toml
[project]
name = "helipad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
